deleted vault/storage keys best-effort afterwards
